ofz#50729 Integer-overflow
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138950
Tested-by: Caolán McNamara <caolanm@redhat.com>

+++ vcl/unx/generic/gdi/cairotextrender.cxx
@@ -190,8 +190,8 @@ void CairoTextRender::DrawTextLayout(const GenericSalLayout& rLayout, const SalG
         return;
 
     const vcl::font::FontSelectPattern& rFSD = rInstance.GetFontSelectPattern();
-    int nHeight = rFSD.mnHeight;
-    int nWidth = rFSD.mnWidth ? rFSD.mnWidth : nHeight;
+    double nHeight = rFSD.mnHeight;
+    double nWidth = rFSD.mnWidth ? rFSD.mnWidth : nHeight;
     if (nWidth == 0 || nHeight == 0)
         return;
